refactor: #565 use transform API in imports-loader
Instead of parsing and doing the AST traversal manually, use the newly created
file transformation feature.

This also should avoid breaking source maps.
